move the task definitions to their own file to segreagate local form remote
